barcode_splitter.py: move classes and functions into new module 'barcodes/splitter.py'.
